Finished adding the remaining code for the game
I added the necessary alerts for the if statements to prompt based on the choice.
I also added console logs to make sure I was getting the right things.
And finally I added the correct condition to check for with the while loop.
The reason this works is because the user will keep getting prompted no matter
how many times they get it wrong since the user value is still not equal to the
generated number. Only when the number guess is exactly equal to the random
number that was generated will we receive the victory alert.