Added button + event listener
Since I found it a bit annoying for the game to always start when you restarted
the page or whenever the page auto restarted after editing code, I decided
to tie the execution of the while loop to a function that would be fired
upon the eventListener listening to a click on the respective button.
This gives users the ability play the game over and over again without
having to restart the page OR without being forced to play the game if
they just want to see what else is on the page.